init-db is able to create postgres dbs and roles
We want the init-db binary to be able to create the dbs and roles
as specified in the config.json file.

This binary would behave as an existing script we had in [1]

We need to use raw string interpolation in most of the queries
as the lib/pq library is unable to prepare statements for
queries like CREATE DATABASE, ALTER ROLE, etc.